+++ hutool-http/src/test/java/cn/hutool/http/server/SimpleServerTest.java
@@ -27,11 +27,13 @@ public class SimpleServerTest {
 				// 文件上传测试
 				// http://localhost:8888/formTest?a=1&a=2&b=3
 				.addAction("/file", (request, response) -> {
-							final UploadFile file = request.getMultipart().getFile("file");
+							final UploadFile[] files = request.getMultipart().getFiles("file");
 							// 传入目录，默认读取HTTP头中的文件名然后创建文件
-							file.write("d:/test/");
-							Console.log("Write file to: d:/test/");
-							response.write(request.getParams().toString(), ContentType.TEXT_PLAIN.toString());
+							for (UploadFile file : files) {
+								file.write("d:/test/");
+								Console.log("Write file: d:/test/" + file.getFileName());
+							}
+							response.write(request.getMultipart().getParamMap().toString(), ContentType.TEXT_PLAIN.toString());
 						}
 				)
 				.start();